requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.0.0=pypi_0
asgiref=3.5.0=pypi_0
astunparse=1.6.3=pypi_0
ca-certificates=2022.3.29=haa95532_1
cached-property=1.5.2=pypi_0
cachetools=5.0.0=pypi_0
certifi=2021.10.8=py37haa95532_2
charset-normalizer=2.0.12=pypi_0
click=8.1.2=pypi_0
colorama=0.4.4=pypi_0
console_shortcut=0.1.1=4
datetime=4.4=pypi_0
django=3.2.13=pypi_0
django-crispy-forms=1.14.0=pypi_0
django-model-utils=4.2.0=pypi_0
filelock=3.6.0=pypi_0
flatbuffers=1.12=pypi_0
gast=0.4.0=pypi_0
google-auth=2.6.6=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.44.0=pypi_0
h5py=3.6.0=pypi_0
huggingface-hub=0.5.1=pypi_0
idna=3.3=pypi_0
importlib-metadata=4.11.3=pypi_0
joblib=1.1.0=pypi_0
keras=2.8.0=pypi_0
keras-nightly=2.10.0.dev2022042807=pypi_0
keras-preprocessing=1.1.2=pypi_0
libclang=14.0.1=pypi_0
markdown=3.3.6=pypi_0
numpy=1.21.6=pypi_0
oauthlib=3.2.0=pypi_0
openssl=1.1.1n=h2bbff1b_0
opt-einsum=3.3.0=pypi_0
packaging=21.3=pypi_0
pandas=1.3.5=pypi_0
pillow=9.1.0=pypi_0
pip=21.2.4=py37haa95532_0
protobuf=3.20.1=pypi_0
psycopg2=2.9.3=pypi_0
psycopg2-binary=2.9.3=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pyparsing=3.0.8=pypi_0
python=3.7.13=h6244533_0
python-dateutil=2.8.2=pypi_0
pytz=2022.1=pypi_0
pyyaml=6.0=pypi_0
regex=2022.4.24=pypi_0
requests=2.27.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
sacremoses=0.0.49=pypi_0
setuptools=61.2.0=py37haa95532_0
six=1.16.0=pypi_0
sqlite=3.38.2=h2bbff1b_0
sqlparse=0.4.2=pypi_0
tb-nightly=2.9.0a20220428=pypi_0
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.8.0=pypi_0
tensorflow-io-gcs-filesystem=0.25.0=pypi_0
termcolor=1.1.0=pypi_0
tf-estimator-nightly=2.10.0.dev2022042808=pypi_0
tf-nightly-gpu=2.10.0.dev20220428=pypi_0
tokenizers=0.12.1=pypi_0
tqdm=4.64.0=pypi_0
transformers=4.18.0=pypi_0
typing-extensions=4.2.0=pypi_0
urllib3=1.26.9=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
werkzeug=2.1.1=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wincertstore=0.2=py37haa95532_2
wrapt=1.14.0=pypi_0
zipp=3.8.0=pypi_0
zope-interface=5.4.0=pypi_0
